Move all extern "C" definitions into lib.rs.
Factoring these out into separate files appears to be insufficient to
force them to all be linked in, replacing libc, so pull them all into
a single module.
