--- c-scape/src/m.rs
@@ -1,798 +1,0 @@
-//! C-compatible `libm` implementation, using the Rust `libm` crate.
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn acos(x: f64) -> f64 {
-    libm::acos(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn acosf(x: f32) -> f32 {
-    libm::acosf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn acosh(x: f64) -> f64 {
-    libm::acosh(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn acoshf(x: f32) -> f32 {
-    libm::acoshf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn asin(x: f64) -> f64 {
-    libm::asin(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn asinf(x: f32) -> f32 {
-    libm::asinf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn asinh(x: f64) -> f64 {
-    libm::asinh(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn asinhf(x: f32) -> f32 {
-    libm::asinhf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn atan(x: f64) -> f64 {
-    libm::atan(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn atan2(x: f64, y: f64) -> f64 {
-    libm::atan2(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn atan2f(x: f32, y: f32) -> f32 {
-    libm::atan2f(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn atanf(x: f32) -> f32 {
-    libm::atanf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn atanh(x: f64) -> f64 {
-    libm::atanh(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn atanhf(x: f32) -> f32 {
-    libm::atanhf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn cbrt(x: f64) -> f64 {
-    libm::cbrt(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn cbrtf(x: f32) -> f32 {
-    libm::cbrtf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn ceil(x: f64) -> f64 {
-    libm::ceil(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn ceilf(x: f32) -> f32 {
-    libm::ceilf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn copysign(x: f64, y: f64) -> f64 {
-    libm::copysign(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn copysignf(x: f32, y: f32) -> f32 {
-    libm::copysignf(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn cos(x: f64) -> f64 {
-    libm::cos(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn cosf(x: f32) -> f32 {
-    libm::cosf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn cosh(x: f64) -> f64 {
-    libm::cosh(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn coshf(x: f32) -> f32 {
-    libm::coshf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn erf(x: f64) -> f64 {
-    libm::erf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn erfc(x: f64) -> f64 {
-    libm::erfc(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn erfcf(x: f32) -> f32 {
-    libm::erfcf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn erff(x: f32) -> f32 {
-    libm::erff(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn exp(x: f64) -> f64 {
-    libm::exp(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn exp2(x: f64) -> f64 {
-    libm::exp2(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn exp2f(x: f32) -> f32 {
-    libm::exp2f(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn exp10(x: f64) -> f64 {
-    libm::exp10(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn exp10f(x: f32) -> f32 {
-    libm::exp10f(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn expf(x: f32) -> f32 {
-    libm::expf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn expm1(x: f64) -> f64 {
-    libm::expm1(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn expm1f(x: f32) -> f32 {
-    libm::expm1f(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn fabs(x: f64) -> f64 {
-    libm::fabs(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn fabsf(x: f32) -> f32 {
-    libm::fabsf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn fdim(x: f64, y: f64) -> f64 {
-    libm::fdim(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn fdimf(x: f32, y: f32) -> f32 {
-    libm::fdimf(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn floor(x: f64) -> f64 {
-    libm::floor(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn floorf(x: f32) -> f32 {
-    libm::floorf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn fma(x: f64, y: f64, z: f64) -> f64 {
-    libm::fma(x, y, z)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn fmaf(x: f32, y: f32, z: f32) -> f32 {
-    libm::fmaf(x, y, z)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn fmax(x: f64, y: f64) -> f64 {
-    libm::fmax(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn fmaxf(x: f32, y: f32) -> f32 {
-    libm::fmaxf(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn fmin(x: f64, y: f64) -> f64 {
-    libm::fmin(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn fminf(x: f32, y: f32) -> f32 {
-    libm::fminf(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn fmod(x: f64, y: f64) -> f64 {
-    libm::fmod(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn fmodf(x: f32, y: f32) -> f32 {
-    libm::fmodf(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn frexp(x: f64, y: *mut i32) -> f64 {
-    let (a, b) = libm::frexp(x);
-    *y = b;
-    a
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn frexpf(x: f32, y: *mut i32) -> f32 {
-    let (a, b) = libm::frexpf(x);
-    *y = b;
-    a
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn hypot(x: f64, y: f64) -> f64 {
-    libm::hypot(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn hypotf(x: f32, y: f32) -> f32 {
-    libm::hypotf(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn ilogb(x: f64) -> i32 {
-    libm::ilogb(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn ilogbf(x: f32) -> i32 {
-    libm::ilogbf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn j0(x: f64) -> f64 {
-    libm::j0(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn j0f(x: f32) -> f32 {
-    libm::j0f(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn j1(x: f64) -> f64 {
-    libm::j1(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn j1f(x: f32) -> f32 {
-    libm::j1f(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn jn(x: i32, y: f64) -> f64 {
-    libm::jn(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn jnf(x: i32, y: f32) -> f32 {
-    libm::jnf(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn ldexp(x: f64, y: i32) -> f64 {
-    libm::ldexp(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn ldexpf(x: f32, y: i32) -> f32 {
-    libm::ldexpf(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn lgamma_r(x: f64, y: *mut i32) -> f64 {
-    let (a, b) = libm::lgamma_r(x);
-    *y = b;
-    a
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn lgammaf_r(x: f32, y: *mut i32) -> f32 {
-    let (a, b) = libm::lgammaf_r(x);
-    *y = b;
-    a
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn log(x: f64) -> f64 {
-    libm::log(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn log1p(x: f64) -> f64 {
-    libm::log1p(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn log1pf(x: f32) -> f32 {
-    libm::log1pf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn log2(x: f64) -> f64 {
-    libm::log2(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn log2f(x: f32) -> f32 {
-    libm::log2f(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn log10(x: f64) -> f64 {
-    libm::log10(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn log10f(x: f32) -> f32 {
-    libm::log10f(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn logf(x: f32) -> f32 {
-    libm::logf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn modf(x: f64, y: *mut f64) -> f64 {
-    let (a, b) = libm::modf(x);
-    *y = b;
-    a
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn modff(x: f32, y: *mut f32) -> f32 {
-    let (a, b) = libm::modff(x);
-    *y = b;
-    a
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn nextafter(x: f64, y: f64) -> f64 {
-    libm::nextafter(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn nextafterf(x: f32, y: f32) -> f32 {
-    libm::nextafterf(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn pow(x: f64, y: f64) -> f64 {
-    libm::pow(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn powf(x: f32, y: f32) -> f32 {
-    libm::powf(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn remainder(x: f64, y: f64) -> f64 {
-    libm::remainder(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn remainderf(x: f32, y: f32) -> f32 {
-    libm::remainderf(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn remquo(x: f64, y: f64, z: *mut i32) -> f64 {
-    let (a, b) = libm::remquo(x, y);
-    *z = b;
-    a
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn remquof(x: f32, y: f32, z: *mut i32) -> f32 {
-    let (a, b) = libm::remquof(x, y);
-    *z = b;
-    a
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn round(x: f64) -> f64 {
-    libm::round(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn roundf(x: f32) -> f32 {
-    libm::roundf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn scalbn(x: f64, y: i32) -> f64 {
-    libm::scalbn(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn scalbnf(x: f32, y: i32) -> f32 {
-    libm::scalbnf(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn sin(x: f64) -> f64 {
-    libm::sin(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn sincos(x: f64, y: *mut f64, z: *mut f64) {
-    let (a, b) = libm::sincos(x);
-    *y = a;
-    *z = b;
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn sincosf(x: f32, y: *mut f32, z: *mut f32) {
-    let (a, b) = libm::sincosf(x);
-    *y = a;
-    *z = b;
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn sinf(x: f32) -> f32 {
-    libm::sinf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn sinh(x: f64) -> f64 {
-    libm::sinh(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn sinhf(x: f32) -> f32 {
-    libm::sinhf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn sqrt(x: f64) -> f64 {
-    libm::sqrt(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn sqrtf(x: f32) -> f32 {
-    libm::sqrtf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn tan(x: f64) -> f64 {
-    libm::tan(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn tanf(x: f32) -> f32 {
-    libm::tanf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn tanh(x: f64) -> f64 {
-    libm::tanh(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn tanhf(x: f32) -> f32 {
-    libm::tanhf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn tgamma(x: f64) -> f64 {
-    libm::tgamma(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn tgammaf(x: f32) -> f32 {
-    libm::tgammaf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn trunc(x: f64) -> f64 {
-    libm::trunc(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn truncf(x: f32) -> f32 {
-    libm::truncf(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn y0(x: f64) -> f64 {
-    libm::y0(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn y0f(x: f32) -> f32 {
-    libm::y0f(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn y1(x: f64) -> f64 {
-    libm::y1(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn y1f(x: f32) -> f32 {
-    libm::y1f(x)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn yn(x: i32, y: f64) -> f64 {
-    libm::yn(x, y)
-}
-
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-unsafe extern "C" fn ynf(x: i32, y: f32) -> f32 {
-    libm::ynf(x, y)
-}
-
-/// Ensure that this module is linked in.
-#[inline(never)]
-#[link_section = ".text.__mustang"]
-#[no_mangle]
-#[cold]
-unsafe extern "C" fn __mustang_c_scape__m() {}
